+++ Editor/PropertyDrawers/InspectOnlyDrawer.cs
@@ -1,0 +1,16 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace Voxell.Inspector
+{
+  [CustomPropertyDrawer(typeof(InspectOnlyAttribute))]
+  public class InspectOnlyDrawer : PropertyDrawer
+  {
+    public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
+    {
+      GUI.enabled = false;
+      EditorGUI.PropertyField(rect, property, label, true);
+      GUI.enabled = true;
+    }
+  }
+}