+++ Resources/python_script.py
@@ -1,3 +1,0 @@
-import UnityEngine as u # type: ignore
-
-u.Debug.Log("Hi from python!")