+++ Runtime/Jobx/Float3MinScanJob.cs
@@ -6,7 +6,7 @@ using Unity.Burst;
 
 namespace Voxell.Jobx
 {
-  public sealed class Float3MinScanJob : Jobx, System.IDisposable
+  public sealed class Float3MinScanJob : System.IDisposable
   {
     private int _valueCount;
     private NativeArray<float3> na_values;
@@ -34,7 +34,7 @@ namespace Voxell.Jobx
       {
         na_prevValues.CopyFrom(na_values);
         minScanJob.offset = offset;
-        jobHandle = minScanJob.Schedule(_valueCount, XL_BATCH_SIZE);
+        jobHandle = minScanJob.Schedule(_valueCount, Jobx.XL_BATCH_SIZE);
         jobHandle.Complete();
       }
       Profiler.EndSample();